moved all store code into components
