+++ src/Components/API/app/routers/hmi.py
@@ -1,0 +1,116 @@
+from fastapi import status, APIRouter
+from bson import ObjectId
+import datetime
+from app import serializers
+from app import schemas
+from app.database import Events, Movements
+import paho.mqtt.publish as publish
+
+
+router = APIRouter()
+
+MQTT_BROKER_URL = "broker.mqttdashboard.com"
+MQTT_BROKER_PORT = 1883
+
+
+@router.get("/events_time", response_description="Get detection events within certain duration")
+def show_event_from_time(start: str, end: str):
+    datetime_start = datetime.datetime.fromtimestamp(int(start), datetime.timezone.utc)
+    datetime_end = datetime.datetime.fromtimestamp(int(end), datetime.timezone.utc)
+    print(datetime_start)
+    print(datetime_end)
+    aggregate = [
+        {
+            '$match':{'timestamp': {'$lt' : datetime_end, '$gte' : datetime_start }}
+            
+        },
+        {
+            '$lookup': {
+                'from': 'species',
+                'localField': 'species',
+                'foreignField': '_id',
+                'as': 'info'
+            }
+        },
+        {
+            "$replaceRoot": { "newRoot": { "$mergeObjects": [ { "$arrayElemAt": [ "$info", 0 ] }, "$$ROOT" ] } }
+        },
+        {
+            '$project': { "audioClip": 0, "sampleRate": 0, "fileFormat": 0}
+        },
+        {
+            "$addFields": {
+            "timestamp": { "$toLong": "$timestamp" }}
+        }       
+
+    ]
+    events = serializers.eventSpeciesListEntity(Events.aggregate(aggregate))
+    return events
+
+@router.get("/audio", response_description="returns audio given ID")
+def show_audio(id: str):
+    aggregate = [
+        {
+            '$match':{'_id': ObjectId(id)}
+        },
+        {
+            '$project': { "audioClip": 1}
+        }
+    ]
+    results = list(Events.aggregate(aggregate))
+    audio = serializers.audioListEntity(results)[0]
+    return audio
+
+@router.get("/movement_time", response_description="Get true animal movement data within certain duration")
+def show_event_from_time(start: str, end: str):
+    datetime_start = datetime.datetime.fromtimestamp(int(start), datetime.timezone.utc)
+    datetime_end = datetime.datetime.fromtimestamp(int(end), datetime.timezone.utc)
+    aggregate = [
+        {
+            '$match':{'timestamp': {'$lt' : datetime_end, '$gte' : datetime_start }}
+        },
+        {
+            '$lookup': {
+                'from': 'species',
+                'localField': 'species',
+                'foreignField': '_id',
+                'as': 'info'
+            }
+        },
+        {
+            "$replaceRoot": { "newRoot": { "$mergeObjects": [ { "$arrayElemAt": [ "$info", 0 ] }, "$$ROOT" ] } }
+        },
+        { "$project": { "info": 0 } },
+        {
+            "$addFields":
+            {
+            "timestamp": { "$toLong": "$timestamp" }
+            }
+        }       
+
+    ]
+    events = serializers.movementSpeciesListEntity(Movements.aggregate(aggregate))
+    return events
+
+@router.get("/microphones", response_description="returns location of all microphones")
+def list_microphones():
+    aggregate = [
+        {
+            "$group":
+            {
+            "_id": "$sensorId",
+            "microphoneLLA": { "$first": "$microphoneLLA" }
+            }
+        }
+    ]
+    results = list(Events.aggregate(aggregate))
+    microphones = serializers.microphoneListEntity(results)
+    return microphones
+
+@router.post("/sim_control", status_code=status.HTTP_201_CREATED)
+def post_control(control: str):
+
+    publish.single("Simulator_Controls", control, hostname=MQTT_BROKER_URL, port=MQTT_BROKER_PORT)
+
+    return control
+
