+++ src/Components/Store/database/sample_data/microphones.json
@@ -1,0 +1,22 @@
+[{
+    "sensorId": "1",
+    "microphoneLLA": [
+        -38.81442552389635,
+        143.5617723719617,
+        23
+    ]
+},{
+    "sensorId": "2",
+    "microphoneLLA": [
+        -38.82713112867141,
+        143.55173018179045,
+        24
+    ]
+},{
+    "sensorId": "3",
+    "microphoneLLA": [
+        -38.826897098553374,
+        143.56447603854625,
+        24
+    ]
+}]
