+++ src/Components/Store/database/sample_data/species.json
@@ -1,0 +1,106 @@
+[{ 
+    "_id": "Sus Scrofa",
+    "commonName": "Wild Pig",
+    "type": "Mammal",
+    "status": "Invasive",
+    "diet": "Omnivore"
+},
+{
+    "_id": "Cervus Unicolour",
+    "commonName": "Sambar Deer",
+    "type": "Mammal",
+    "status": "Invasive",
+    "diet": "Herbavore"
+},
+{
+    "_id": "Aegotheles Cristatus",
+    "commonName": "Australian Owlet-nightjar",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Carnivore"
+},
+{
+    "_id": "Alauda Arvensis",
+    "commonName": "European Skylark",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Omnivore"
+},
+{
+    "_id": "Caligavis Chrysops",
+    "commonName": "Yellow-faced Honeyeater",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Omnivore"
+},
+{
+    "_id": "Capra Hircus",
+    "commonName": "Feral Goat",
+    "type": "Mammal",
+    "status": "Invasive",
+    "diet": "Herbavore"
+},
+{
+    "_id": "Colluricincla Harmonica",
+    "commonName": "Grey Shrikethrush",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Omnivore"
+},
+{
+    "_id": "Corvus Coronoides",
+    "commonName": "Australian Raven",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Omnivore"
+},
+{
+    "_id": "Dama Dama",
+    "commonName": "Fallow Deer",
+    "type": "Mammal",
+    "status": "Invasive",
+    "diet": "Herbavore"
+},
+{
+    "_id": "Eopsaltria Australis",
+    "commonName": "Eastern Yellow Robin",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Carnivore"
+},
+{
+    "_id": "Felis Catus",
+    "commonName": "Cat",
+    "type": "Mammal",
+    "status": "Invasive",
+    "diet": "Carnivore"
+},
+{
+    "_id": "Pachycephala Rufiventris",
+    "commonName": "Rufous Whistler",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Omnivore"
+},
+{
+    "_id": "Ptilotula Penicillata",
+    "commonName": "White-plumed Honeyeater",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Omnivore"
+},
+{
+    "_id": "Rattus Norvegicus",
+    "commonName": "Brown Rat",
+    "type": "Mammal",
+    "status": "Invasive",
+    "diet": "Omnivore"
+},
+{
+    "_id": "Strepera Graculina",
+    "commonName": "Pied Currawong",
+    "type": "Bird",
+    "status": "Least Concern",
+    "diet": "Omnivore"
+}
+]
