+++ src/Components/Store/otway_species.txt
@@ -1,0 +1,133 @@
+Oriolus sagittatus
+Antigone rubicunda
+Myiagra inquieta
+Accipiter fasciatus
+Phaps chalcoptera
+Eopsaltria australis
+Hieraaetus morphnoides
+Biziura lobata
+Artamus superciliosus
+Acanthiza chrysorrhoa
+Colluricincla harmonica
+Cracticus torquatus
+Acanthorhynchus tenuirostris
+Petrochelidon nigricans
+Coracina novaehollandiae
+Falco berigora
+Gallinula tenebrosa
+Anthochaera carunculata
+Falco peregrinus
+Daphoenositta chrysoptera
+Pachycephala rufiventris
+Aquila audax
+Malurus cyaneus
+Eurostopodus mystacalis
+Manorina melanocephala
+Caligavis chrysops
+Cygnus atratus
+Anhinga novaehollandiae
+Corvus coronoides
+Aythya australis
+Gavicalis virescens
+Petroica rosea
+Callocephalon fimbriatum
+Myzomela sanguinolenta
+Falco cenchroides
+Macropus giganteus
+Sericornis frontalis
+Lathamus discolor
+Ptilotula penicillata
+Cisticola exilis
+Pachycephala pectoralis
+Spilopelia chinensis
+Anthochaera chrysoptera
+Myiagra rubecula
+Trichosurus vulpecula
+Turdus merula
+Ninox boobook
+Rhipidura rufifrons
+Tachybaptus novaehollandiae
+Artamus cyanopterus
+Elanus axillaris
+Ptilonorhynchus violaceus
+Limnodynastes tasmaniensis
+Psephotus haematonotus
+Phylidonyris novaehollandiae
+Acanthiza nana
+Dacelo novaeguineae
+Austronomus australis
+Grallina cyanoleuca
+Ocyphaps lophotes
+Notamacropus rufogriseus
+Lalage tricolor
+Milvus migrans
+Acanthiza reguloides
+Chalcites basalis
+Petroica phoenicea
+Vulpes vulpes
+Corvus mellori
+Ninox strenua
+Gerygone olivacea
+Cacatua sanguinea
+Ceyx azureus
+Microeca fascinans
+Phylidonyris pyrrhopterus
+Limnodynastes dumerilii
+Hirundo neoxena
+Petroica boodang
+Corcorax melanorhamphos
+Eudynamys orientalis
+Turnix varius
+Phaps elegans
+Cormobates leucophaea
+Rhipidura leucophrys
+Acanthiza pusilla
+Melithreptus lunatus
+Zosterops lateralis
+Dicaeum hirundinaceum
+Aegotheles cristatus
+Glossopsitta concinna
+Anas superciliosa
+Petrochelidon ariel
+Acanthiza lineata
+Zoothera lunulata
+Chalcites lucidus
+Nycticorax caledonicus
+Neochmia temporalis
+Cacomantis variolosus
+Rhipidura albiscapa
+Pardalotus punctatus
+Strepera versicolor
+Limnodynastes peronii
+Cacomantis flabelliformis
+Anseranas semipalmata
+Podargus strigoides
+Coturnix pectoralis
+Vanellus miles
+Fulica atra
+Cacatua galerita
+Pelecanus conspicillatus
+Alisterus scapularis
+Stagonopleura guttata
+Cinclosoma punctatum
+Egretta novaehollandiae
+Elseyornis melanops
+Acanthagenys rufogularis
+Ninox connivens
+Myiagra cyanoleuca
+Pseudocheirus peregrinus
+Carduelis carduelis
+Merops ornatus
+Melithreptus brevirostris
+Pardalotus striatus
+Chenonetta jubata
+Wallabia bicolor
+Strepera graculina
+Accipiter cirrocephalus
+Eolophus roseicapilla
+Falcunculus frontatus
+Acrocephalus australis
+Platycercus elegans
+Crinia signifera
+Todiramphus sanctus
+Haliastur sphenurus
